refactor: move jsonpath helpter to ct_apitools--helper.inc.php
